day difference is now Long

+++ app/src/main/java/de/fhe/ai/weemeal/app/ui/screens/core/AppNavigationHost.kt
@@ -119,7 +119,7 @@ fun AppNavigationHost(
             Screen.AddRecipeToWeekList.route,
             Screen.AddRecipeToWeekList.navigationCommand(0).arguments
         ) { entry ->
-            val cookingDateDaysAhead = entry.arguments?.getInt("cookingDateDaysAhead")
+            val cookingDateDaysAhead = entry.arguments?.getLong("cookingDateDaysAhead")
             val vm by viewModel<AddRecipeToWeekListViewModel>(
                 parameters = {
                     parametersOf(cookingDateDaysAhead)
